backend/requirements.txt
Flask==2.2.2
Flask-Script==2.0.6
python-dotenv==0.20.0
pandas==1.4.3
beautifulsoup4==4.11.1
gunicorn==20.1.0
odfpy==1.4.1